feat: improv error handling etc
