+++ src/app.html
@@ -45,7 +45,7 @@
       referrerpolicy="no-referrer" />
     <link rel="icon" type="image/png" href="/favicon.png?v=3" />
     <link rel="apple-touch-icon" href="/favicon.png?v=3" />
-    <link rel="stylesheet" href="/global.css?v=25052023" type="text/css" />
+    <link rel="stylesheet" href="/global.css?v=26052023" type="text/css" />
 
     <style lang="postcss">
       html {
