+++ src/routes/[game]/[asset]/+page.server.js
@@ -6,6 +6,7 @@ export async function load({ params }) {
     const res = await fetch(`https://api.wanderer.moe/game/${game}/${asset}`);
     if (res.ok) {
       const response = await res.json();
+      if (response.error) throw error(404, "API Error or Page Not Found");
       return {
         response,
         game,
