+++ static/global.css
@@ -60,7 +60,6 @@ a {
   z-index: 999;
 }
 
-
 .popupcontainer {
   position: fixed;
   top: 50%;
